move check_missing_divergent tests to traits/mod, fix paths, add to CI
