+++ refinery/src/traits/mod.rs
@@ -84,3 +84,152 @@ pub(crate) const ASSERT_MIGRATIONS_TABLE: &str =
 
 pub(crate) const GET_APPLIED_MIGRATIONS: &str = "SELECT version, name, applied_on, checksum \
      FROM refinery_schema_history ORDER BY version ASC;";
+
+#[cfg(test)]
+mod tests {
+    use super::{check_missing_divergent, AppliedMigration, Error, Migration};
+
+    fn get_migrations() -> Vec<Migration> {
+        let migration1 = Migration::from_filename(
+            "V1__initial.sql",
+            include_str!("../../tests/sql_migrations/V1-2/V1__initial.sql"),
+        )
+        .unwrap();
+
+        let migration2 = Migration::from_filename(
+            "V2__add_cars_and_motos_table.sql",
+            include_str!("../../tests/sql_migrations/V1-2/V2__add_cars_and_motos_table.sql"),
+        )
+        .unwrap();
+
+        let migration3 = Migration::from_filename(
+            "V3__add_brand_to_cars_table",
+            include_str!("../../tests/sql_migrations/V3/V3__add_brand_to_cars_table.sql"),
+        )
+        .unwrap();
+
+        let migration4 = Migration::from_filename(
+            "V4__add_year_field_to_cars",
+            &"ALTER TABLE cars ADD year INTEGER;",
+        )
+        .unwrap();
+
+        vec![migration1, migration2, migration3, migration4]
+    }
+
+    #[test]
+    fn check_missing_divergent_returns_all_migrations_if_applied_are_empty() {
+        let migrations = get_migrations();
+        let applied: Vec<AppliedMigration> = Vec::new();
+        let result = check_missing_divergent(applied, migrations.clone(), true, true).unwrap();
+        assert_eq!(migrations, result);
+    }
+
+    #[test]
+    fn check_missing_divergent_returns_unapplied() {
+        let migrations = get_migrations();
+        let applied: Vec<AppliedMigration> = vec![
+            migrations[0].clone().as_applied(),
+            migrations[1].clone().as_applied(),
+            migrations[2].clone().as_applied(),
+        ];
+        let remaining = vec![migrations[3].clone()];
+        let result = check_missing_divergent(applied, migrations, true, true).unwrap();
+        assert_eq!(remaining, result);
+    }
+
+    #[test]
+    fn check_missing_divergent_fails_on_divergent() {
+        let migrations = get_migrations();
+        let mut applied: Vec<AppliedMigration> = vec![
+            migrations[0].clone().as_applied(),
+            migrations[1].clone().as_applied(),
+            migrations[2].clone().as_applied(),
+        ];
+        applied[2].checksum = "3a6d3a3".into();
+        let migration = migrations[2].clone();
+        let err = check_missing_divergent(applied, migrations, true, true).unwrap_err();
+        match err {
+            Error::DivergentVersion(applied, divergent) => {
+                assert_eq!(migration, divergent);
+                assert_eq!("add_brand_to_cars_table", applied.name);
+            }
+            _ => panic!("failed test"),
+        }
+    }
+
+    #[test]
+    fn check_missing_divergent_doesnt_fail_on_divergent() {
+        let migrations = get_migrations();
+        let mut applied: Vec<AppliedMigration> = vec![
+            migrations[0].clone().as_applied(),
+            migrations[1].clone().as_applied(),
+            migrations[2].clone().as_applied(),
+        ];
+        applied[2].checksum = "3a6d3a3".into();
+        let remaining = vec![migrations[3].clone()];
+        let result = check_missing_divergent(applied, migrations, false, true).unwrap();
+        assert_eq!(remaining, result);
+    }
+
+    #[test]
+    fn check_missing_divergent_fails_on_missing_on_applied() {
+        let migrations = get_migrations();
+        let applied: Vec<AppliedMigration> = vec![
+            migrations[0].clone().as_applied(),
+            migrations[2].clone().as_applied(),
+        ];
+        let migration = migrations[1].clone();
+        let err = check_missing_divergent(applied, migrations, true, true).unwrap_err();
+        match err {
+            Error::MissingVersion(missing) => {
+                assert_eq!(migration.as_applied(), missing);
+            }
+            _ => panic!("failed test"),
+        }
+    }
+
+    #[test]
+    fn check_missing_divergent_fails_on_missing_on_filesystem() {
+        let mut migrations = get_migrations();
+        let applied: Vec<AppliedMigration> = vec![
+            migrations[0].clone().as_applied(),
+            migrations[1].clone().as_applied(),
+            migrations[2].clone().as_applied(),
+        ];
+        let migration = migrations.remove(1);
+        let err = check_missing_divergent(applied, migrations, true, true).unwrap_err();
+        match err {
+            Error::MissingVersion(missing) => {
+                assert_eq!(migration.as_applied(), missing);
+            }
+            _ => panic!("failed test"),
+        }
+    }
+
+    #[test]
+    fn check_missing_divergent_doesnt_fail_on_missing_on_applied() {
+        let migrations = get_migrations();
+        let applied: Vec<AppliedMigration> = vec![
+            migrations[0].clone().as_applied(),
+            migrations[2].clone().as_applied(),
+        ];
+        let remaining = vec![migrations[3].clone()];
+        let result = check_missing_divergent(applied, migrations, true, false).unwrap();
+        assert_eq!(remaining, result);
+    }
+
+    #[test]
+    fn check_missing_divergent_doesnt_fail_on_missing_on_filesystem() {
+        let mut migrations = get_migrations();
+        let applied: Vec<AppliedMigration> = vec![
+            migrations[0].clone().as_applied(),
+            migrations[1].clone().as_applied(),
+            migrations[2].clone().as_applied(),
+        ];
+        migrations.remove(1);
+        let remaining = vec![migrations[2].clone()];
+        let result = check_missing_divergent(applied, migrations, true, false).unwrap();
+        assert_eq!(remaining, result);
+    }
+}
